CHROMIUM: drm/bridge: anx7625: Fix variable is not used correctly
Fixes: da7f49e653bb6 ("CHROMIUM: drm/bridge: anx7625: Add DSC feature support")

Forcefully convert int-type variable into a 64-bit variable.

BUG=b:399512650
BRANCH=none
TEST=tested on navi
UPSTREAM-TASK=b:399512650

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6333372
Tested-by: Pengjian Lu <lupengjian@huaqin.corp-partner.google.com>
Commit-Queue: Fei Shao <fshao@chromium.org>

+++ drivers/gpu/drm/bridge/analogix/anx7625.c
@@ -2578,7 +2578,7 @@ static bool anx7625_bridge_mode_fixup(struct drm_bridge *bridge,
 		adj_hsync = DSC_HSYNC_LEN;
 		adj_hfp = DSC_HFP_LEN;
 		adj_hbp = DSC_HBP_LEN;
-		vref = (u32)div_u64(adj->clock * 1000 * 1000,
+		vref = (u32)div_u64((u64)adj->clock * 1000 * 1000,
 				    adj->htotal * adj->vtotal);
 		goto calculate_timing;
 	}
@@ -2668,7 +2668,7 @@ calculate_timing:
 	adj->hsync_end = adj->hsync_start + adj_hsync;
 	adj->htotal = adj->hsync_end + adj_hbp;
 	if (mode->clock > DSC_PIXEL_CLOCK)
-		adj->clock = (u32)div_u64(vref * adj->htotal * adj->vtotal,
+		adj->clock = (u32)div_u64((u64)vref * adj->htotal * adj->vtotal,
 					  1000 * 1000);
 
 	DRM_DEV_DEBUG_DRIVER(dev, "hsync_start(%d), hsync_end(%d), htot(%d)\n",
